+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Update the MNIST example to be easier to read (#208)
 * Add How to download a model on the substrafl example (#204)
 
 ## [0.22.0]