+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add How to download a model on the substrafl example (#204)
 - Update colors & ux (#209)
+- Update the MNIST example to be easier to read (#208)
 
 ## [0.22.0]