+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add How to download a model on the substrafl example (#204)
 - Update colors & ux (#209)
 - Update the MNIST example to be easier to read (#208)
+- Move utils folder content in assets folder (#210)
 
 ## [0.22.0]