+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increase tocdepth in SDK API reference to display methods under Client (#219)
 - Update substratools Docker image name (#222)
 - Use the new `add_metric` SubstraFL function in examples (#214)
+- Add documentation on mTLS setup (#200)
 
 ## [0.22.0]