+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.23.0]
+## [0.23.1]
 
 - Add Mnist example image (#218)
 - Add a Sickit-learn example for SubstraFL to the documentation (#206)